Upgrade to v16 tg bot lib

+++ musicallychallenged/Domain/ActiveContestEntry.cs
@@ -10,7 +10,7 @@ namespace musicallychallenged.Domain
         [Key]
         public int Id { get; set; }
 
-        public int AuthorUserId { get; set; }
+        public long AuthorUserId { get; set; }
         public int ChallengeRoundNumber { get; set; }
 
         public Instant Timestamp { get; set; }
